Null validation

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -19,6 +20,10 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
 
       // retrieve domain entity object
       var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
+      if (leaveTypeToDelete == null) 
+      {
+         throw new NotFoundException(nameof(LeaveType), request.Id);
+      }
 
       // remove from database
       await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
